changes total into float

+++ db/migrate/20160616202516_add_total_to_bills.rb
@@ -1,5 +1,5 @@
 class AddTotalToBills < ActiveRecord::Migration
   def change
-    add_column :bills, :total, :integer
+    add_column :bills, :total, :float
   end
 end
